+++ protocols/relay/src/lib.rs
@@ -47,15 +47,12 @@ pub mod inbound {
     pub mod hop {
         pub use crate::protocol::inbound_hop::FatalUpgradeError;
     }
-    pub mod stop {
-        pub use crate::protocol::inbound_stop::FatalUpgradeError;
-    }
 }
 
 /// Types related to the relay protocol outbound.
 pub mod outbound {
     pub mod hop {
-        pub use crate::protocol::outbound_hop::FatalUpgradeError;
+        pub use crate::protocol::outbound_hop::{ConnectError, ProtocolViolation, ReserveError};
     }
     pub mod stop {
         pub use crate::protocol::outbound_stop::FatalUpgradeError;